+++ rust/cubestore/cubestore/src/remotefs/queue.rs
@@ -1,11 +1,9 @@
 use crate::config::ConfigObj;
 use crate::di_service;
-use crate::metastore::MetaStore;
 use crate::remotefs::{RemoteFile, RemoteFs};
 use crate::util::lock::acquire_lock;
-use crate::{app_metrics, CubeError};
+use crate::CubeError;
 use async_trait::async_trait;
-use chrono::Utc;
 use core::fmt;
 use datafusion::cube_ext;
 use deadqueue::unlimited;
@@ -15,15 +13,12 @@ use smallvec::alloc::fmt::Formatter;
 use std::collections::HashSet;
 use std::fmt::Debug;
 use std::fs::Metadata;
-use std::path::Path;
 use std::sync::Arc;
 use tokio::sync::{broadcast, watch, RwLock};
-use tokio::time::Duration;
 
 pub struct QueueRemoteFs {
     config: Arc<dyn ConfigObj>,
     remote_fs: Arc<dyn RemoteFs>,
-    metastore: Arc<dyn MetaStore>,
     upload_queue: unlimited::Queue<RemoteFsOp>,
     download_queue: unlimited::Queue<RemoteFsOp>,
     // TODO not used
@@ -63,17 +58,12 @@ pub enum RemoteFsOpResult {
 di_service!(QueueRemoteFs, [RemoteFs]);
 
 impl QueueRemoteFs {
-    pub fn new(
-        config: Arc<dyn ConfigObj>,
-        remote_fs: Arc<dyn RemoteFs>,
-        metastore: Arc<dyn MetaStore>,
-    ) -> Arc<Self> {
+    pub fn new(config: Arc<dyn ConfigObj>, remote_fs: Arc<dyn RemoteFs>) -> Arc<Self> {
         let (stopped_tx, stopped_rx) = watch::channel(false);
         let (tx, rx) = broadcast::channel(16384);
         Arc::new(Self {
             config,
             remote_fs,
-            metastore,
             upload_queue: unlimited::Queue::new(),
             download_queue: unlimited::Queue::new(),
             deleted: RwLock::new(HashSet::new()),
@@ -135,29 +125,10 @@ impl QueueRemoteFs {
             }));
         }
 
-        let to_move = queue_remote_fs.clone();
-        if queue_remote_fs.config.upload_to_remote() {
-            futures.push(tokio::task::spawn(async move {
-                to_move.cleanup_local_files_loop().await;
-            }));
-        }
         join_all(futures)
             .await
             .into_iter()
             .collect::<Result<Vec<_>, _>>()?;
-        Ok(())
-    }
-
-    #[allow(dead_code)]
-    pub async fn wait_remote_fs_cleanup_loop(
-        queue_remote_fs: Arc<Self>,
-        stat_tags: Vec<String>,
-    ) -> Result<(), CubeError> {
-        let to_move = queue_remote_fs.clone();
-        tokio::task::spawn(async move {
-            to_move.cleanup_remotefs_loop(stat_tags).await;
-        })
-        .await?;
         Ok(())
     }
 
@@ -210,193 +181,6 @@ impl QueueRemoteFs {
             x => panic!("Unexpected operation: {:?}", x),
         }
         Ok(())
-    }
-
-    async fn cleanup_local_files_loop(&self) {
-        let local_dir = self.local_path().await;
-        let mut stopped_rx = self.stopped_rx.clone();
-        let cleanup_interval = Duration::from_secs(self.config.local_files_cleanup_interval_secs());
-        let cleanup_local_files_delay =
-            Duration::from_secs(self.config.local_files_cleanup_delay_secs());
-
-        loop {
-            // Do the cleanup every now and then.
-            tokio::select! {
-                () = tokio::time::sleep(cleanup_interval) => {},
-                res = stopped_rx.changed() => {
-                    if res.is_err() || *stopped_rx.borrow() {
-                        return;
-                    }
-                }
-            }
-
-            let local_dir_copy = local_dir.clone();
-            let res_local_files =
-                cube_ext::spawn_blocking(move || -> Result<HashSet<String>, std::io::Error> {
-                    let mut local_files = HashSet::new();
-                    for res_entry in Path::new(&local_dir_copy).read_dir()? {
-                        let entry = match res_entry {
-                            Err(_) => continue, // ignore errors, might come from concurrent fs ops.
-                            Ok(e) => e,
-                        };
-
-                        let ft = match entry.file_type() {
-                            Err(_) => continue,
-                            Ok(ft) => ft,
-                        };
-                        if !ft.is_file() {
-                            continue;
-                        }
-
-                        let file_name = match entry.file_name().into_string() {
-                            Err(_) => {
-                                log::error!("could not convert file name {:?}", entry.file_name());
-                                continue;
-                            }
-                            Ok(name) => name,
-                        };
-
-                        if !file_name.ends_with(".parquet") {
-                            continue;
-                        }
-
-                        let should_deleted = if let Ok(metadata) = entry.metadata() {
-                            match metadata.created() {
-                                Ok(created) => {
-                                    if created
-                                        .elapsed()
-                                        .map_or(true, |e| e < cleanup_local_files_delay)
-                                    {
-                                        false
-                                    } else {
-                                        true
-                                    }
-                                }
-                                Err(e) => {
-                                    log::error!(
-                                        "error while getting created time for file {:?}:{}",
-                                        entry.file_name(),
-                                        e
-                                    );
-                                    false
-                                }
-                            }
-                        } else {
-                            false
-                        };
-                        if !should_deleted {
-                            continue;
-                        }
-
-                        local_files.insert(file_name);
-                    }
-                    Ok(local_files)
-                })
-                .await
-                .unwrap();
-
-            let mut local_files = match res_local_files {
-                Err(e) => {
-                    log::error!("error while trying to list local files: {}", e);
-                    continue;
-                }
-                Ok(f) => f,
-            };
-
-            if local_files.is_empty() {
-                continue;
-            }
-
-            let files_from_metastore = match self.metastore.get_all_filenames().await {
-                Err(e) => {
-                    log::error!("could not get the list of files from metastore: {}", e);
-                    continue;
-                }
-                Ok(f) => f,
-            };
-
-            // Only keep the files we want to remove in `local_files`.
-            for f in files_from_metastore {
-                local_files.remove(&f);
-            }
-
-            if !local_files.is_empty() {
-                log::debug!(
-                    "Cleaning up {} files that were removed remotely",
-                    local_files.len()
-                );
-                log::trace!("The files being removed are {:?}", local_files);
-            }
-
-            let local_dir_copy = local_dir.clone();
-            cube_ext::spawn_blocking(move || {
-                for f in local_files {
-                    let _ = std::fs::remove_file(Path::new(&local_dir_copy).join(f));
-                }
-            })
-            .await
-            .unwrap();
-        }
-    }
-
-    async fn cleanup_remotefs_loop(&self, stat_tags: Vec<String>) -> () {
-        let mut stopped_rx = self.stopped_rx.clone();
-        let cleanup_interval =
-            Duration::from_secs(self.config.remote_files_cleanup_interval_secs());
-        let cleanup_local_files_delay = self.config.remote_files_cleanup_delay_secs() as i64;
-        loop {
-            // Do the cleanup every now and then.
-            tokio::select! {
-                () = tokio::time::sleep(cleanup_interval) => {},
-                res = stopped_rx.changed() => {
-                    if res.is_err() || *stopped_rx.borrow() {
-                        return;
-                    }
-                }
-            }
-
-            let res_remote_files = self.list_with_metadata("").await;
-            let remote_files = match res_remote_files {
-                Err(e) => {
-                    log::error!("could not get the list of remote files: {}", e);
-                    continue;
-                }
-                Ok(f) => f,
-            };
-
-            let files_from_metastore = match self.metastore.get_all_filenames().await {
-                Err(e) => {
-                    log::error!("could not get the list of files from metastore: {}", e);
-                    continue;
-                }
-                Ok(f) => f.into_iter().collect::<HashSet<_>>(),
-            };
-
-            let mut files_to_remove = Vec::new();
-            let mut files_to_remove_size = 0;
-
-            for f in remote_files {
-                if files_from_metastore.get(f.remote_path()).is_some() {
-                    continue;
-                }
-                if Utc::now()
-                    .signed_duration_since(f.updated().clone())
-                    .num_seconds()
-                    < cleanup_local_files_delay
-                {
-                    continue;
-                }
-                files_to_remove.push(f.remote_path().to_string());
-                files_to_remove_size += f.file_size;
-            }
-            if !files_to_remove.is_empty() {
-                app_metrics::REMOTE_FS_FILES_TO_REMOVE
-                    .report_with_tags(files_to_remove.len() as i64, Some(&stat_tags));
-
-                app_metrics::REMOTE_FS_FILES_SIZE_TO_REMOVE
-                    .report_with_tags(files_to_remove_size as i64, Some(&stat_tags));
-            }
-        }
     }
 }
 
@@ -681,65 +465,6 @@ mod test {
         path
     }
     #[tokio::test]
-    async fn queue_cleanup_local_files() {
-        Config::test("cleanup_local_files")
-            .update_config(|mut c| {
-                c.local_files_cleanup_delay_secs = 2;
-                c.local_files_cleanup_interval_secs = 1;
-                c
-            })
-            .start_test(async move |services| {
-                let service = services.sql_service;
-                let meta_store = services.meta_store;
-                let remote_fs = services.injector.get_service_typed::<dyn RemoteFs>().await;
-                let _ = service.exec_query("CREATE SCHEMA test").await.unwrap();
-                let _ = service
-                    .exec_query("CREATE TABLE test.tst (a int, b int)")
-                    .await
-                    .unwrap();
-                let _ = service
-                    .exec_query("INSERT INTO test.tst (a, b) VALUES (10, 10), (20 , 20)")
-                    .await
-                    .unwrap();
-                let _ = service
-                    .exec_query("INSERT INTO test.tst (a, b) VALUES (20, 20), (40 , 40)")
-                    .await
-                    .unwrap();
-                let files = meta_store.get_all_filenames().await.unwrap();
-                assert_eq!(files.len(), 2);
-                for f in files.iter() {
-                    let path = remote_fs.local_file(&f).await.unwrap();
-                    assert!(Path::new(&path).exists());
-                }
-                let path = remote_fs.local_file("metastore").await.unwrap();
-                assert!(Path::new(&path).exists());
-
-                meta_store
-                    .delete_chunks_without_checks(vec![1])
-                    .await
-                    .unwrap();
-
-                assert_eq!(meta_store.get_all_filenames().await.unwrap().len(), 1);
-                for f in files.iter() {
-                    let path = remote_fs.local_file(&f).await.unwrap();
-                    assert!(Path::new(&path).exists());
-                }
-                Delay::new(Duration::from_millis(3000)).await; // TODO logger init conflict
-
-                let path = remote_fs.local_file(&files[0]).await.unwrap();
-                assert!(!Path::new(&path).exists());
-
-                let path = remote_fs.local_file(&files[1]).await.unwrap();
-                assert!(Path::new(&path).exists());
-
-                let path = remote_fs.local_file("metastore").await.unwrap();
-                assert!(Path::new(&path).exists());
-
-                let _ = service.exec_query("SELECT * FROM test.tst").await.unwrap();
-            })
-            .await;
-    }
-    #[tokio::test]
     async fn queue_upload() {
         let config = Config::test("upload_retries_all_fail");
         config.configure_injector().await;
@@ -748,11 +473,7 @@ mod test {
             error: MockFSError::None,
             download_error: false,
         });
-        let queue_fs = QueueRemoteFs::new(
-            config.config_obj(),
-            failed_fs.clone(),
-            config.injector().get_service_typed().await,
-        );
+        let queue_fs = QueueRemoteFs::new(config.config_obj(), failed_fs.clone());
 
         let path = make_test_csv();
 
@@ -775,11 +496,7 @@ mod test {
             error: MockFSError::WrongSize,
             download_error: false,
         });
-        let queue_fs = QueueRemoteFs::new(
-            config.config_obj(),
-            failed_fs.clone(),
-            config.injector().get_service_typed().await,
-        );
+        let queue_fs = QueueRemoteFs::new(config.config_obj(), failed_fs.clone());
 
         let path = make_test_csv();
 
@@ -802,11 +519,7 @@ mod test {
             error: MockFSError::MissingFile,
             download_error: false,
         });
-        let queue_fs = QueueRemoteFs::new(
-            config.config_obj(),
-            failed_fs.clone(),
-            config.injector().get_service_typed().await,
-        );
+        let queue_fs = QueueRemoteFs::new(config.config_obj(), failed_fs.clone());
 
         let path = make_test_csv();
 
@@ -828,7 +541,6 @@ mod test {
         let queue_fs = QueueRemoteFs::new(
             config.config_obj(),
             config.injector().get_service("original_remote_fs").await,
-            config.injector().get_service_typed().await,
         );
         let r = tokio::spawn(QueueRemoteFs::wait_processing_loops(queue_fs.clone()));
         let res = queue_fs.download_file("temp-upload/foo.csv", None).await;
@@ -850,7 +562,6 @@ mod test {
         let queue_fs = QueueRemoteFs::new(
             config.config_obj(),
             config.injector().get_service("original_remote_fs").await,
-            config.injector().get_service_typed().await,
         );
         let r = tokio::spawn(QueueRemoteFs::wait_processing_loops(queue_fs.clone()));
         queue_fs
@@ -882,11 +593,7 @@ mod test {
             download_error: true,
         });
         let path = make_test_csv();
-        let queue_fs = QueueRemoteFs::new(
-            config.config_obj(),
-            failed_fs.clone(),
-            config.injector().get_service_typed().await,
-        );
+        let queue_fs = QueueRemoteFs::new(config.config_obj(), failed_fs.clone());
 
         let r = tokio::spawn(QueueRemoteFs::wait_processing_loops(queue_fs.clone()));
